+++ spine-unity/Assets/Spine Examples/Scripts/RaggedySpineboy.cs
@@ -38,11 +38,11 @@ namespace Spine.Unity.Examples {
 		public float restoreDuration = 0.5f;
 		public Vector2 launchVelocity = new Vector2(50,100);
 
-		Spine.Unity.Modules.SkeletonRagdoll2D ragdoll;
+		Spine.Unity.Examples.SkeletonRagdoll2D ragdoll;
 		Collider2D naturalCollider;
 
 		void Start () {
-			ragdoll = GetComponent<Spine.Unity.Modules.SkeletonRagdoll2D>();
+			ragdoll = GetComponent<Spine.Unity.Examples.SkeletonRagdoll2D>();
 			naturalCollider = GetComponent<Collider2D>();
 		}