+++ spine-unity/Assets/Spine Examples/Scripts/Sample Components/Legacy/AtlasRegionAttacher.cs
@@ -30,9 +30,9 @@
 using UnityEngine;
 using System.Collections.Generic;
 using Spine;
-using Spine.Unity.Modules.AttachmentTools;
+using Spine.Unity.AttachmentTools;
 
-namespace Spine.Unity.Modules {
+namespace Spine.Unity.Examples {
 	/// <summary>
 	/// Example code for a component that replaces the default attachment of a slot with an image from a Spine atlas.</summary>
 	public class AtlasRegionAttacher : MonoBehaviour {