+++ spine-unity/Assets/Spine Examples/Scripts/Sample Components/SkeletonUtility Modules/Editor/SkeletonRagdollInspector.cs
@@ -32,7 +32,7 @@
 using UnityEngine;
 using UnityEditor;
 
-namespace Spine.Unity.Modules {
+namespace Spine.Unity.Examples {
 	
 	public class SkeletonRagdollInspector : UnityEditor.Editor {
 		[CustomPropertyDrawer(typeof(SkeletonRagdoll.LayerFieldAttribute))]